feat(derived-divine_rarity): add Deriveddivine_rarity with unique property and method
